requirements.txt
Flask
Flask-Session
psycopg2
SQLAlchemy
argon2_cffi
